+++ belt-hash/src/lib.rs
@@ -10,12 +10,12 @@
 
 pub use digest::{self, Digest};
 
-mod block_api;
-pub use block_api::{BeltHashCore, belt_compress};
+/// Block-level types
+pub mod block_api;
 
 digest::buffer_fixed!(
     /// BelT hasher state.
-    pub struct BeltHash(BeltHashCore);
+    pub struct BeltHash(block_api::BeltHashCore);
     oid: "1.2.112.0.2.0.34.101.31.81";
     impl: FixedHashTraits;
 );