+++ gost94/src/lib.rs
@@ -10,15 +10,16 @@
 
 pub use digest::{self, Digest};
 
-mod block_api;
+/// Block-level types
+pub mod block_api;
+/// GOST94 parameters.
 pub mod params;
 
-pub use block_api::Gost94Core;
 use params::Gost94Params;
 
 digest::buffer_fixed!(
     /// GOST94 hash function generic over parameters.
-    pub struct Gost94<P: Gost94Params>(Gost94Core<P>);
+    pub struct Gost94<P: Gost94Params>(block_api::Gost94Core<P>);
     impl: FixedHashTraits;
 );