+++ gost94/src/params.rs
@@ -1,5 +1,3 @@
-//! GOST94 parameters.
-
 #[cfg(feature = "oid")]
 use digest::const_oid::{AssociatedOid, ObjectIdentifier};