+++ jh/src/lib.rs
@@ -9,11 +9,10 @@
 
 pub use digest::{self, Digest};
 
-mod block_api;
+/// Block-level types
+pub mod block_api;
 mod compressor;
 mod consts;
-
-pub use block_api::JhCore;
 
 use digest::{
     block_api::CtOutWrapper,
@@ -22,21 +21,21 @@ use digest::{
 
 digest::buffer_fixed!(
     /// JH-224 hasher.
-    pub struct Jh224(CtOutWrapper<JhCore, U28>);
+    pub struct Jh224(CtOutWrapper<block_api::JhCore, U28>);
     impl: BaseFixedTraits Default Clone HashMarker Reset FixedOutputReset;
 );
 digest::buffer_fixed!(
     /// JH-256 hasher.
-    pub struct Jh256(CtOutWrapper<JhCore, U32>);
+    pub struct Jh256(CtOutWrapper<block_api::JhCore, U32>);
     impl: BaseFixedTraits Default Clone HashMarker Reset FixedOutputReset;
 );
 digest::buffer_fixed!(
     /// JH-384 hasher.
-    pub struct Jh384(CtOutWrapper<JhCore, U48>);
+    pub struct Jh384(CtOutWrapper<block_api::JhCore, U48>);
     impl: BaseFixedTraits Default Clone HashMarker Reset FixedOutputReset;
 );
 digest::buffer_fixed!(
     /// JH-512 hasher.
-    pub struct Jh512(CtOutWrapper<JhCore, U64>);
+    pub struct Jh512(CtOutWrapper<block_api::JhCore, U64>);
     impl: BaseFixedTraits Default Clone HashMarker Reset FixedOutputReset;
 );