+++ sha1/src/block_api.rs
@@ -13,7 +13,7 @@ use digest::{
 #[cfg(feature = "zeroize")]
 use digest::zeroize::{Zeroize, ZeroizeOnDrop};
 
-pub use crate::compress;
+pub use crate::compress::compress;
 
 const STATE_LEN: usize = 5;
 const H0: [u32; STATE_LEN] = [0x67452301, 0xEFCDAB89, 0x98BADCFE, 0x10325476, 0xC3D2E1F0];