+++ sha2/src/lib.rs
@@ -25,50 +25,47 @@ use digest::{
     consts::{U28, U32, U48, U64},
 };
 
+/// Block-level types
+pub mod block_api;
+
 #[rustfmt::skip]
 mod consts;
-mod block_api;
 mod sha256;
 mod sha512;
 
-pub use sha256::compress256;
-pub use sha512::compress512;
-
-pub use block_api::{Sha256VarCore, Sha512VarCore};
-
 digest::buffer_fixed!(
     /// SHA-256 hasher.
-    pub struct Sha256(CtOutWrapper<Sha256VarCore, U32>);
+    pub struct Sha256(CtOutWrapper<block_api::Sha256VarCore, U32>);
     oid: "2.16.840.1.101.3.4.2.1";
     impl: FixedHashTraits;
 );
 digest::buffer_fixed!(
     /// SHA-384 hasher.
-    pub struct Sha384(CtOutWrapper<Sha512VarCore, U48>);
+    pub struct Sha384(CtOutWrapper<block_api::Sha512VarCore, U48>);
     oid: "2.16.840.1.101.3.4.2.2";
     impl: FixedHashTraits;
 );
 digest::buffer_fixed!(
     /// SHA-512 hasher.
-    pub struct Sha512(CtOutWrapper<Sha512VarCore, U64>);
+    pub struct Sha512(CtOutWrapper<block_api::Sha512VarCore, U64>);
     oid: "2.16.840.1.101.3.4.2.3";
     impl: FixedHashTraits;
 );
 digest::buffer_fixed!(
     /// SHA-224 hasher.
-    pub struct Sha224(CtOutWrapper<Sha256VarCore, U28>);
+    pub struct Sha224(CtOutWrapper<block_api::Sha256VarCore, U28>);
     oid: "2.16.840.1.101.3.4.2.4";
     impl: FixedHashTraits;
 );
 digest::buffer_fixed!(
     /// SHA-512/224 hasher.
-    pub struct Sha512_224(CtOutWrapper<Sha512VarCore, U28>);
+    pub struct Sha512_224(CtOutWrapper<block_api::Sha512VarCore, U28>);
     oid: "2.16.840.1.101.3.4.2.5";
     impl: FixedHashTraits;
 );
 digest::buffer_fixed!(
     /// SHA-512/256 hasher.
-    pub struct Sha512_256(CtOutWrapper<Sha512VarCore, U32>);
+    pub struct Sha512_256(CtOutWrapper<block_api::Sha512VarCore, U32>);
     oid: "2.16.840.1.101.3.4.2.6";
     impl: FixedHashTraits;
 );