+++ sha3/src/block_api.rs
@@ -1,11 +1,11 @@
-use crate::{DEFAULT_ROUND_COUNT, PLEN, Sha3ReaderCore};
+use crate::{DEFAULT_ROUND_COUNT, PLEN};
 use core::{fmt, marker::PhantomData};
 use digest::{
     HashMarker, Output,
     array::ArraySize,
     block_api::{
         AlgorithmName, Block, BlockSizeUser, Buffer, BufferKindUser, Eager, ExtendableOutputCore,
-        FixedOutputCore, OutputSizeUser, Reset, UpdateCore,
+        FixedOutputCore, OutputSizeUser, Reset, UpdateCore, XofReaderCore,
     },
     crypto_common::{
         BlockSizes,
@@ -13,6 +13,8 @@ use digest::{
     },
     typenum::{IsLessOrEqual, True, U0, U200},
 };
+
+pub use crate::cshake::{CShake128Core, CShake256Core};
 
 /// Core Sha3 fixed output hasher state.
 #[derive(Clone)]
@@ -229,6 +231,78 @@ where
             _pd: PhantomData,
         })
     }
+}
+
+/// Core Sha3 XOF reader.
+#[derive(Clone)]
+pub struct Sha3ReaderCore<Rate, const ROUNDS: usize = DEFAULT_ROUND_COUNT>
+where
+    Rate: BlockSizes + IsLessOrEqual<U200, Output = True>,
+{
+    state: [u64; PLEN],
+    _pd: PhantomData<Rate>,
+}
+
+impl<Rate, const ROUNDS: usize> Sha3ReaderCore<Rate, ROUNDS>
+where
+    Rate: BlockSizes + IsLessOrEqual<U200, Output = True>,
+{
+    pub(crate) fn new(state: &[u64; PLEN]) -> Self {
+        Self {
+            state: *state,
+            _pd: PhantomData,
+        }
+    }
+}
+
+impl<Rate, const ROUNDS: usize> BlockSizeUser for Sha3ReaderCore<Rate, ROUNDS>
+where
+    Rate: BlockSizes + IsLessOrEqual<U200, Output = True>,
+{
+    type BlockSize = Rate;
+}
+
+impl<Rate, const ROUNDS: usize> XofReaderCore for Sha3ReaderCore<Rate, ROUNDS>
+where
+    Rate: BlockSizes + IsLessOrEqual<U200, Output = True>,
+{
+    #[inline]
+    fn read_block(&mut self) -> Block<Self> {
+        let mut block = Block::<Self>::default();
+        for (src, dst) in self.state.iter().zip(block.chunks_mut(8)) {
+            dst.copy_from_slice(&src.to_le_bytes()[..dst.len()]);
+        }
+        keccak::p1600(&mut self.state, ROUNDS);
+        block
+    }
+}
+
+impl<Rate, const ROUNDS: usize> Drop for Sha3ReaderCore<Rate, ROUNDS>
+where
+    Rate: BlockSizes + IsLessOrEqual<U200, Output = True>,
+{
+    fn drop(&mut self) {
+        #[cfg(feature = "zeroize")]
+        {
+            use digest::zeroize::Zeroize;
+            self.state.zeroize();
+        }
+    }
+}
+
+impl<Rate, const ROUNDS: usize> fmt::Debug for Sha3ReaderCore<Rate, ROUNDS>
+where
+    Rate: BlockSizes + IsLessOrEqual<U200, Output = True>,
+{
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.write_str("Sha3ReaderCore { ... }")
+    }
+}
+
+#[cfg(feature = "zeroize")]
+impl<Rate, const ROUNDS: usize> digest::zeroize::ZeroizeOnDrop for Sha3ReaderCore<Rate, ROUNDS> where
+    Rate: BlockSizes + IsLessOrEqual<U200, Output = True>
+{
 }
 
 pub(crate) fn xor_block(state: &mut [u64; PLEN], block: &[u8]) {